requirements.txt
# Core
Django>=4.2,<5.2
python-dotenv>=1.0.0

# Database — PostgreSQL
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9

# WebSocket / ASGI
# channels 4.1+ fixes async_unsafe bugs in AuthMiddlewareStack
channels>=4.1.0
daphne>=4.0.0
asgiref>=3.7.0

# NLP
textblob>=0.18.0
nltk>=3.8.0
requests>=2.28.0

# Auth / Media
Pillow>=10.0.0

# AI
groq>=0.4.0

# Production
django-cors-headers>=4.3.0
whitenoise>=6.6.0
gunicorn>=21.2.0
